Implement MediaCapabilitiesDatabase with LevelDB
The LevelDB implementation supports the production use case of storing
video decode statistics for recall to answer MediaCapabilities queries.

This CL implements the interface and tests. A later CL will instantiate
the DB as part of the VideoDecodePerfHistory service for actual usage.

Kudos to mlamouri@ for first implementing this in
https://chromium-review.googlesource.com/c/chromium/src/+/625962

TEST: added to media_unittests
Bug: 695264

+++ media/mojo/services/media_capabilities_database.h
@@ -39,9 +39,9 @@ class MEDIA_MOJO_EXPORT MediaCapabilitiesDatabase {
 
   // Information saves to identify the capabilities related to a given |Entry|.
   struct MEDIA_MOJO_EXPORT Info {
-    Info(uint32_t frames_decoded, uint32_t frames_dropped);
-    uint32_t frames_decoded;
-    uint32_t frames_dropped;
+    Info(uint64_t frames_decoded, uint64_t frames_dropped);
+    uint64_t frames_decoded;
+    uint64_t frames_dropped;
   };
 
   virtual ~MediaCapabilitiesDatabase() {}
